Temporary fix for compiling Belvedere.exe.
On my Windows 7 64-Bit machine, I am unable to compile directly with Compile_AHK. As a workaround, I'm compiling first
with ahk2exe, and then with Compile_AHK. This seems to have to side effect of not setting the program icon properly.
I'm going to test this on 32-bit version of Windows to see if it makes any difference.